Speed up Playlist lookup. Resolves #10
Populating the selection list is much faster now.

The slow lookup process has just been moved to the background of the
downloader instead. That means the actual downloading will take the same
amount of time as before. This can't really be "fixed", youtube-dl
requires some time to acquire the needed information for each video.

But to make this faster, it will lookup the required information in the
background and download whatever videos it currently has the required
info for.